modified:   CHANGELOG.md modified:   README.md new file:   doc/bot-preview-post.png deleted:    doc/instagram-profile.png modified:   src/bot.py modified:   src/constants.py renamed:    src/tools/__init__.py -> src/modules/__init__.py renamed:    src/tools/downloader.py -> src/modules/downloader.py renamed:    src/tools/uploader.py -> src/modules/uploader.py
